changed how permissions worked
-moved permissions to separate activity in order to have access to permissions before the map is loaded on main activity
-added a current location marker and reset map position button on map
