+++ app/src/main/res/layout/map_fragment.xml
@@ -9,6 +9,6 @@
 
     <com.google.android.gms.maps.MapView
         android:id="@+id/map_view"
-        android:layout_width="412dp"
-        android:layout_height="798dp" />
+        android:layout_width="match_parent"
+        android:layout_height="match_parent" />
 </LinearLayout>
